+++ packages/paperclip-utils/src/core/constants.ts
