+++ packages/paperclip-utils/src/core/memo.ts
