+++ packages/paperclip-utils/src/core/utils.ts
