+++ packages/paperclip-utils/src/css/exports.ts
@@ -5,22 +5,6 @@ export type CSSExports = {
   keyframes: Record<string, KeyframesExport>;
 };
 
-export type PCExports = {
-  components: Record<string, ComponentExport>;
-  style: CSSExports;
-};
-
-export type ComponentProperty = {
-  name: string;
-  optional: boolean;
-};
-
-export type ComponentExport = {
-  name: string;
-  properties: Record<string, ComponentProperty>;
-  public: boolean;
-};
-
 export type ClassNameExport = {
   name: string;
   scopedName: string;
