+++ packages/paperclip-utils/src/html/tree.ts
@@ -1,4 +1,4 @@
-import { memoize } from "./memo";
+import { memoize } from "../core/memo";
 import { nodePathToAry } from "./virt";
 
 // core tree utils
