+++ packages/paperclip-utils/src/index.ts
@@ -1,29 +1,29 @@
-export * from "./events";
-export * from "./virt";
-export * from "./ast";
-export * from "./js-ast";
-export * from "./actions";
-export * from "./js-virt";
-export * from "./stringify-sheet";
-export * from "./css-ast";
-export * from "./base-ast";
-export * from "./config";
-export * from "./constants";
+export * from "./core/events";
+export * from "./html/virt";
+export * from "./html/ast";
+export * from "./script/ast";
+export * from "./core/actions";
+export * from "./script/virt";
+export * from "./css/stringify-sheet";
+export * from "./css/ast";
+export * from "./base/ast";
+export * from "./core/config";
+export * from "./core/constants";
 // export * from "./errors";
-export * from "./graph";
-export * from "./virt-mtuation";
-export * from "./resolve";
-export * from "./stringify-virt-node";
-export * from "./css-virt";
-export * from "./virt-patcher";
-export * from "./exports";
-export * from "./utils";
-export * from "./source-watcher";
-export * from "./virt-css-patcher";
-export * from "./memo";
-export * from "./base-virt";
-export * from "./errors";
-export * from "./diagnostics";
-export * from "./tree";
-export * from "./style-inspection";
-export * from "./module-ast";
+export * from "./core/graph";
+export * from "./html/virt-mtuation";
+export * from "./core/resolve";
+export * from "./html/stringify-virt-node";
+export * from "./css/virt";
+export * from "./html/patcher";
+export * from "./core/utils";
+export * from "./core/source-watcher";
+export * from "./css/patcher";
+export * from "./core/memo";
+export * from "./base/virt";
+export * from "./core/errors";
+export * from "./core/diagnostics";
+export * from "./html/tree";
+export * from "./css/style-inspection";
+export * from "./core/module-ast";
+export * from "./css/exports";
